requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
sqlalchemy==2.0.48
python-dotenv==1.2.2
cryptography==46.0.6
pyopenssl==26.0.0
reportlab==4.4.10
google-genai==1.68.0
transformers==5.4.0
torch==2.11.0
spacy==3.8.13
